Refactor locale module into multiple files
